ROM loader/scanner bug fixes

+++ Core/burner/cocoa/FXZipFile.h
@@ -27,7 +27,7 @@
 }
 
 @property (nonatomic, strong) NSString *filename;
-@property (nonatomic, assign) NSUInteger CRC;
+@property (nonatomic, assign) UInt32 CRC;
 @property (nonatomic, assign) NSUInteger length;
 
 @end
